+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [0.0.2] - 2019-04-02
+### Added
+- is_remote:  Determines if the remote git repository exists.
 
 
 ## [0.0.1] - 2019-03-19